Only configure git user if needed
There are multiple git behaviors linked to valid user configuration;
one among them is the --signoff flag and/or `format.signOff`
configuration. Always setting the user to `rpm-build <rpm-build>` can
then cause git to emit nonsense lines, which can not be easily turned
off just for the affected repositories.

Setting the user only if one is not already configured should give
best of both worlds – git will not bail out on unknown user
and the actual author of commits will be kept in Signed-off-By trailers.

Signed-off-by: Jan Staněk <jstanek@redhat.com>